Refactored TrackMultipleValues out.
